move misplaced import logic to mathesar namespace
